move allocation functions to their own .py
