+++ flowsa/data_source_scripts/Blackhurst_IO.py
@@ -13,8 +13,8 @@ import pandas as pd
 import numpy as np
 from flowsa.common import US_FIPS
 from flowsa.flowbyfunctions import assign_fips_location_system, \
-    proportional_allocation_by_location_and_activity, \
     load_fba_w_standardized_units
+from flowsa.allocation import proportional_allocation_by_location_and_activity
 from flowsa.mapping import add_sectors_to_flowbyactivity
 from flowsa.data_source_scripts.BLS_QCEW import clean_bls_qcew_fba
 from flowsa.validation import compare_df_units
